+++ src/main/kotlin/in/kkkev/jjidea/jj/cli/TemplateParts.kt
@@ -12,5 +12,10 @@ object TemplateParts {
         return "${fn("commit_id")} ++ \"~\" ++ ${fn("commit_id")}.shortest()"
     }
 
+    fun nameWithRemote(base: String? = null): String {
+        fun fn(name: String) = fn(base, name)
+        return "if(${fn("remote")}, ${fn("name")} ++ \"@\" ++ ${fn("remote")}, ${fn("name")})"
+    }
+
     private fun fn(base: String?, name: String) = base?.let { "$it.$name()" } ?: name
 }